double -> float conversions for sun safe related modules

+++ src/architecture/msgPayloadDefC/THRArrayOnTimeCmdMsgPayload.h
@@ -25,7 +25,7 @@
 
 /*! @brief Structure used to define the output definition for vehicle effectors*/
 typedef struct {
-    double OnTimeRequest[MAX_EFF_CNT];     //!< - Control request fraction array
+    float OnTimeRequest[MAX_EFF_CNT];     //!< - Control request fraction array
 }THRArrayOnTimeCmdMsgPayload;
